refactor after merging master into branch
